New test on computing the counts_alive and counts_blocked lists at the end of the simulation process
Currently the test fails for some simulations generated with reactivate=True, but passes for all tried simulation setups when ractivate=False.
Note that the `finalize_type` parameter currently has no effect on the calculations of counts as the simulation is NOT finalized prior to computing the said lists of counts.