First working version of the simpler simulation implementation that simulates ALL the particles at the beginning (following my handwritten notes)
This implementation allows for the correct reactivation, namely: the absorbed particle is reactivated to another particle number by reading its position *at precisely the absorption time t0* (which was not the case in the previous implementation, where the reactivated position could occur at a time < t0.

This implementation in addition copies all the relevant events from the assigned particle for reactivation (Q) to the reactivated particle P, which was *not* the case in the previous implementation either.

The implementation is tested through a new test called test_simpler_algorithm() in test_QB.py.