Fixed a subtle bug related to retrieving incorrect particle numbers associated to particle IDs when useing FinalizeType.REMOVE_CENSORED
The particle number P was incorrectly assigned to an incorrect value when a removed particle r was among the N particles present in the system, as in that case the 'particle number' attribute was set to None and the value of P was set equal to p, which is wrong if p refers to a particle whose ID is larger than the removed particle r but still smaller than N-1. In fact that p will no longer correspond to particle *number* p because there is one less element in the self.info_particles list after the removal of particle r!

This bug is subtle because it only happens when the above conditions are satisfied which in turn only happen when we run a very small number of iterations on a large number of particles N, making it more likely that particle among the first N particles is removed when finalizing the simulation with FinalizeType.REMOVE_CENSORED.

The bug was fixed by simply setting the particle number when initializing the self.info_particles list as this is already known upfront!