[QB] Working version of the Fleming Viot system - Approx. 1, including code clean up from non used attributes and methods
One important realization leading to the working version is that the conditional probability of blocking (P(K / T>t, s=1)) is computed SIMPLY from the empirical position distribution of the particles (m(K,xi)), i.e. we DO NOT NEED TO CONDITION on T>t, s=1 to compute that probabiliy!!! Amazing...
This implies that we do not need to copy all the past history of the assigned particle to the reactivated particle, making the process much much faster! (as the number of relevant events stored in the self.info_particles list is much much smaller).