[TD] Implemented analysis of effective learning rate alpha
The effective learning rate is the learning rate alpha multiplied by the eligibility trace of the state, which is the effective learning rate that is being applied in TD(lambda) learning.

Notes:
- Detail about learning rate being used for states at the TD(lambda) update:
I noticed that the learning rate alpha applied for all states at every learning sttep in TD(lambda) is defined by the alpha of the VISITED state at the time of the update, as opposed to the alpha of the corresponding state that is affected by the update (recall that ALL states are updated in TD(lambda), not only the state visited last). For more details see the note in the code.

Updates in td.py:
- The effective alpha is now computed as alpha * eligibility_trace_z(s) for deeper analysis of the correct implementation of TD(lambda) and its adaptive flavour.
- New function _PlotAlphasEffective() that plots the effective alphas for selected states chosen among the middle ones.
- Implemented the option (currently commented out) of a first-visit TD(lambda) (as opposed to every-visit TD(lambda)), following the lecture notes by Shimkin on TD(lambda).