[QB] Adaptation to multi-server system #1: succeeded in making the generate_one_iteration() work for now
It was crucial to make the Getters in QueueMM class and GeneicQueue class return a COPY of the class' attribute when the attribute is NOT a scalar.
For instance, this is the case now with the `size` attribute of the GenericQueue class that stores the queue size for each server.
If we don't do this copy, any updates of the array attributes of the QueueMM class will ALSO affect the array objects where the returned value by the Getter was stored (as in e.g. `last_event_times = self.particles[P].getTimeLastEvent()`, which made the event times of the particle trajectories be ALL the same!)