[QB] Solved the bias observed by a slow convergence of the FV estimator of the blocking probability
The solution was to use an estimation of the survival probability P(T>t / s=1) from an N=1 particle simulation running LONG enough (i.e. from a Monte-Carlo estimation obtained with a
reactivate=False call to the simulation process).

Reason: the estimation of the survival probability P(T>t / s=1) is not affected by the number of particles, but rather by the simulation time, because it requires the killing event to happen, and for this we need time, NOT particles.

In order to implement this, I added a new parameter to the EstimatorQueueBlockingFlemingViot constructor that receives an estimate of the survival probability as a pandas Dataframe object.

New results PDF files were created with the results obtained for the convergence analysis.

DETAILS:
- estimators.py:
  - constructor: new parameter `proba_survival_given_activation` which is a pandas Dataframe with two columns: the times at which the probability is measured and the survival probability itself.
  - compute_survival_probability_from_counts(): new parameter `t` which receives the list of times at which the counts used to compute the survival probability are measured. Otherwise, the process failed when called from the outside world.
  - estimate_proba_survival_and_blocking_conditional(): implemented the estimation of the survival probability given activation, and the blocking probability with common time axis using the list of survival probability values given by the user.
  - insert_relative_time(): now an warning is issued when a negative relative time would be inserted (as opposed to `break`ing the process).

- test/test_QB.py:
  - test_fv_implementation():
    - the constructor used in the FV estimation now receives the survival probability given activation estimated by the MC run.
    - added a plot that compares the estimation of the survival probability curve (P(T>t / s=1)) between MC and FV, i.e. between N=1 particle with long run and N>>1 particle with short run.