[FV] New queue environment receiving a GenericQueue object in environments/queues.py
Goal: Be able to remove the duplication of the queue information that is stored and updated during the FV simulation by passing just this new queue environment to the estimator defined in the EstimatorQueueBlockingFlemingViot class (as opposed to passing the queue environment AND the GenericQueue or QueueMM object, which implies that both objects storing the queue need to be updated when a new job arrives or a new job is served).

Test #3 in estimators.py was updated with the new signature of the EnvQueueSingleBufferWithJobClasses class.