[FV] Implemented estimation of survival curve and E(T) based on N trajectories run until FIRST ABSORPTION
IMPORTANT: The estimation of E(T) based on N trajectories until first absorption is WRONG, as explained by Matt in a 2-hour call today, 27-Apr-2021. In fact, the estimation of E(T) should be based on the duration of cycles of return to the start state observed in the SAME trajectory. We CANNOT start the trajectory all over again in order to estimate E(T) once the particle is absorbed, because the stationary distribution of the returning states is NOT the same as the stationary distribution of the states in the original process.
For more information, see Asmussen, pag. 170 when he talks about Regeneration processes.

DETAILS:
- estimators.py:
  - run_simulation(): New parameter stop_at_first_absorption=False which indicates whether the simulation of the queue should be stopped when the first absorption occurs. This is used to estimate the survival curve P(T>t) more reliably by running N particles until the first absorption instead of running just one trajectory and recording every pass by the activation state. The reason the latter approach is less reliable is that the number of samples collected in that case may be very little, especially when the activation state is very close to K (because the stationary probability that the system is at such a high state is very low).

  - Added recording of number of samples used to estimate each measure of interest: MC-based blocking probability, survival curve, expected return time to absorption.

- test_QB.py: adapted the code to the new setup of the estimators.py functions and class methods as described above.