[FV] test_QB.py: improved plots vs. buffer_size_activation by using proportions
Goal: More easily interpret the plots, so that buffer sizes for
activation increase from left to right even when some of the activation
sizes are given as integral numbers (as opposed to proportions).

Also added a plot vs. the integral buffer size for activation, besides
the one against its proportion.