[FV] Initial implementation of random selection of START state folowing a distribution
Goal: Simulate the multi-server scenario where the absorption and activation positions have more than one configuration for the Markov state (e.g. (1,0,0), (0,1,0), (0,0,1) all correspond to a buffer size = 1).

The random selection of states is implemented in the new method simulate_expected_survival_time() in estimators.py.
There is also a unit test on the method added at the bottom of the file.