[FV] Fixed SIMULATION by eliminating generation of death times for IDLE servers (Ref: call on Fri, 07-May-2021)
NOTE: It's not clear whether this version will work, because it was created by hand-picking the changes that have to do with the elimination of death times generation on idle servers.

NOTE: On Mon, 10-May-2021 in the morning we agreed on a call with Urtzi that the simulation I am doing is trajectorial (as opposed to Markovian), i.e. I am doing the simulation deemed more difficult by Matt (compared to the Markovian simulation) because we need to keep track of the time of occurrence of the state changes. However, I didn't find it particularly difficult because I am storing all generated next event times as ABSOLUTE times! (which avoids the program having to update the remaining service time or the remaining time for the next arrival whenever the system is in a new state.