[FV] Implemented a method that generates next birth times for each server DIRECTLY
Goal: Reduce the number of possible source of errors in the simulation, by eliminating the layer where we simulate arriving job classes and only THEN do we assign the new job to a server.

NOTE: It's not clear whether this version will work, because it was created by hand-picking the changes that have to do with this commit.