[FV] New functions to check the simulated system (estimated lambdas and mus)
Goal: Make sure that the simulated system reflects what we think we are simulating.
This was indeed verified as the estimated lambdas, mus, and rhos are close enough to the true lambdas and mus with which the system is simulated.
The expected absorption time E(T) was also verified for absorption set s=0 in the single-server system (which is equal to 1/Pr(0) + 1/lambda, where 1/lambda is the average sojourn time in s=0).
However, the value of E(T) for the absorption set s=(0,0,0) was NOT verified for the multi-server system, and we don't know yet whether there is something wrong with the implementation (it doesn't seem so, as we verified with Urtzi and Matt that the program is correctly measuring the cycle time defined by consecutive absorptions.

DETAILS:
- estimators.py: New functions:
  - compute_observed_rates()
  - check_expected_survival_time()