[FV] Changed the simulation setup to have P(T>t) estimated from the FV simulation
Goal: Make the estimation of P(T>t) and Phi(t) be based on the SAME start state. This is important in the multi-server system because there are multiple possible states for the same BSA value. The estimation of the blocking probability should be based on the SAME start state for both P(T>t) and Phi(t).

NOTE HOWEVER that, although this change removed the positive bias on the estimated blocking probability by FV, it is not really the correct implementation, because, as discussed today with Matt, we should be starting each simulation on the basis of the exit distribution from the absorption set. This is NOT the stationary distribution restricted to activation set and should be estimated from an initial Monte-Carlo simulation.

Note also that, after this change, the following was the impact on the estimated blocking probability:
- the variance was reduced VERY considerably, BUT...
- a NEGATIVE BIAS is now observed in certain circumstances, mostly likely due to an underestimation of the P(T>t) because of not enough simulation time (which is actually why the estimation of P(T>t) from a separate Monte-Carlo simulation came into play in the first place). E.g. for K=40, BSA=0.5K or BSA=0.25K. The bias is NOT observed for K=20, BSA=0.5K though, and this is because (i) K is not so large; (ii) the activation set is NOT so far away from K.

A NEGATIVE BIAS is also observed in the single-server case now! And this is because P(T>t) is underestimated because of not enough simulation time --as mentioned above.

Note in particular that this negative bias (e.g. for single-server K=20) is reduced when we get closer to K, but this is because E(T) is now ALSO underestimated and its underestimation compensates the underestimation of P(T>t)!! (So, not a really good thing...)