[FV] Changes committed in Toshiba on 14-May-2021 + changes stashed as WIP + changes in MSI
This commit came about as follows:
- On 13-May-2021 I copied the repository from Toshiba to MSI via an USB
in order to run simulations in MSI. The status of the repository at the
time of the copy was as follows: it had all the changes that would be later commited in Toshiba on
14-May-2021 + the changes done afterwards that would be later stashed as WIP in Toshiba.

- I did a few changes in MSI, mostly regarding the simulation process in test_QB.py, such as
improving the plot generation process, and improving the generated plots
to be included in the paper to RLNQ-2021.

DETAILS: In test_QB.py
- New function plot_estimates1() which produces just ONE plot of the
estimated values (as opposed to a 1x2 plot layout, done by
plot_estimates()).

- Fixed the problem with using the same seed for the estimation of
P(T>t) and E(T) which was creating phantom correlations!

- Moved the generation of the plots in the analyze_estimates() function
so that they are generated AFTER the results are saved (so that an error
in their generation does not make us lose all the results!!

- Implemented the save of the plots to a file (via plt.savefig()).