[FV] New functions in estimators.py that compute the distribution of event times
The original function name was check_expected_survival_time() and it was used only to compute the distribution of times to absorption, but now the function is called plot_event_times_dist() and can be used to compute the distribution of the time to any valid special event (EventType).

The computation of the true expected time to absorption (survival time) is now done by the new function compute_true_expected_survival_time().