[FV] Fixed the simulation using the indirect assignment of jobs to servers
Both the MC and the FV simulations for the single-server case now give the same results whether they are run:
- by direct assignment of jobs to servers (using the equivalent job arrival rate) --> using method self._generate_birth_times_from_equivalent_rates()
- by indirect assignment of jobs to servers, i.e. via the assignment policy --> using method self._generate_birth_times()

In fact, the results of these two simulations should NOT differ as there is only ONE server to which jobs can be assigned!

The fix was achieved by the following changes:
- for the MC simulation: by eliminating a call to np.random.choice() when the arriving job classes are assigned to servers when nservers = 1 --> This guaranteed that the exponentially simulated times for the next birth and death events are the same either when we use self._generate_birth_times_from_equivalent_rates() or self.__generate_birth_times().
- for the FV simulation: by resetting the object attributes that are used by self._generate_birth_times() to generate the next birth times. This reset is done in the self._create_new_particles() method.

Note that 3 tests have been put in place (with assertions) in estimators.py for both the single server and the multi-server cases for the MC and FV simulations: 2.1, 2.2 for single server, and 3.1 for multi-server.
In the multi-server case, two different assertion sets have been written, one for the results obtained with the direct assignment of jobs to servers and one for the results obtained with the indirect assignment of jobs (via the assignment policy).
Note that in the multi-server case, the results of each approach are different because there is a random component in the random assignment of jobs to servers, as jobs arrive.

In addition:
- Refactored the _generate_birth_times() method by defining auxiliary functions that carry out the three main steps: generate next times by job class, accept jobs, assign accepted jobs to servers. This increases clarity of the method.
- Function plot_event_times_dist() now also returns the number of observations used in the calculation of E(T).