[FV] Added log file and results from the long-run simulation run at IRIT cluster which took 10 days
The log file was downloaded via Filezilla to my laptop from the IRIT cluster c8ni.irit.fr and then parsed with parse_log.awk to generate the CSV file with the results, since this was not generated due to the fact that the simulation process stopped because of an assertion error.

The variance and bias were analyzed in the Excel file that had already been created for this purpose. Unfortunately, I could not analyze the variance of the integral because this information is not output to the log file...

I also slightly modified the original Excel with the analysis of variance and bias for the simulation I had run on my laptop.