[FV] test_QB.py: new boolean parameter in plot_results_fv_mc() that avoids fitting a spline to the points
Goal: avoid the error that "m > k must hold" generated by the spline-fitting function which apparently happens when we have too few points to fit.
(Ref: https://stackoverflow.com/questions/57621728/scipy-interpolate-2d-error-m-kx1ky1-must-hold)

Prepared the test run #9 (used for K=60) to the new setup of all-in-one estimation with limited burn-in simulation time.

Note that:
- I used N=6400, the same as the number of particles used in the all-in-one estimation with all N particles required to be used in the estimation of E(T) and P(T>t) that took 15 days to complete.
At the same time this number is LARGER than N=1000, the value used for K=40 because we need more particles to get a few particles blocking. We can do a calculation for this that is based on the fact that the probability of blocking under FV is ~ sqrt(K) rho^(K/2), with which we should be able to say that it is ~ sqrt(K) * sqrt(Pr(K))
- I used nmeantimes=1000, the SAME value as for K=40, because we would like to be able to compare with the case K=40.