[FV] All-in-one estimation with limited time: new results added for #servers=1 & #servers=3 (performed second week of Sep-2021)
IMPORTANT: See note on commit 38be6eb about the management of censored times in this implementation for the estimation of E(T) and P(T>t) and the limitations it entails.
NEXT: Implement a separate estimation of E(T) and P(T>t) which is based on a single particle let run for a long time, say N*Tsim.

This All-in-one estimation uses a limited fixed time for the burn-in period that is used to estimate E(T) and P(T>t), as opposed to using ALL N particles for ALL activation sizes J considered.

This setup shows the trade-off between a better estimation of E(T) and P(T>t) for small J (as they are based on a larger number of particles) and a good estimation of Phi(t) for large J (as blocking occurs with higher frequency).