[FV] Updated tests results for tests 2.1, 2.2, 3.1 to match the All-in-one estimation process on limited time
The only result to update was the *estimated blocking probability by FV*.

Although these results should have not changed after the new implementation, I verified that the change is due to different trajectories being taken by the particles due to a different order of the random number generation. in fact, the plots of the usual blue curve (P(T>t)) and red curve (Phi(t)) are *completely different* between this version and the version where the tests pass, namely the one with commit cafb2fc, which is the version just prior to the implementation of the reduced simulation time in the All-in-one estimation process.

Note that, to ease the check of test in future versions, I added the number of events in all cases when reporting the estimated blocking probability (this was already in place in test 3.1 and it helped).