[Learners] mc.py: made active unit tests pass
At first unit tests didn't pass because the two terminal states in the 1D gridworld had an expected value of 0.0 while at some point I had changed them to the more natural values of -1.0 and +1.0 for the left and right terminal states, respectively.

The process also failed because of the np.printoptions() function which is not available for the older numpy version 1.14.0 which is the one installed at IRIT's cluster. So I changed how the numpy print options are set.

Still one of the unit tests in test_MC_lambda.py fails (the last one) and it is not easy to find out what the problem is.