[FV-RL] Updated the unit tests in simulators.py following the changes in MC and FV listed below
The changes done recently that called for the update of the unit tests are:
- MC:
  - Fixed the estimation of the stationary probability by considering the continuous-time observations instead of the discrete-time observations. This largely decreased the stationary probability (e.g. from 0.87 to 0.31, which is actually closer to the true probability of K-1, when K = 2).
  - Changed the start state of the simulation from J to J-1.

- FV:
  - Fixed the estimation of P(T>t) and E(T_A) by applying the parameterized policy (instead of blocking at K, which was the case before, when calling the EstimatorQueueBlockingFlemingViot class).

Note that the above changes affect  both the IGA learning mode and the REINFORCE_TRUE learning mode. IGA might be *surprisingly* affected, but recall that that method still looks into the values of Pr(K-1) and Pr(K) and computes grad(V), even if these quantities are not needed for the update of theta.