[TD] test_TD_lambda.py: updated expected results in 5 of 6 tests to make them pass
The only test that already passed was the first test which actually includes 6 data_provider tests.

Out of the 6 tests, 3 are of the TD(lambda) adaptive case. In all these adaptive algorithms, the value of lambda is set as the Boltzmann function, 1 - exp(-|x|), where x = delta(t) / average( abs(V(t)) ).

The other parameters leading to the given expected results are listed just before setting the value of `expected_values`.