[TD] Fixed the use of the learning rate alpha in the TD learners when lambda > 0
The alpha value was set to the same value for ALL states visited in the past, namely to the alpha calculated for the *currently* visited state. This is conceptually wrong but was used because I observed that learning was faster in the gridworld environments. However, this may not be the case in other environments, and in fact a non-convergence was more easily observed in e.g. the Mountain Car environment.

DETAILS:
- td.py:
  - learn_pred_V() in both classes: Fixed the learning rate alpha used to update the value function of states visited in the past when lambda > 0: before this change, I was using the learning rate alpha set for the *currently visited state* for ALL past states.

- test_TD_lambda.py:
    - Updated the expected results of the tests with lambda > 0 and adaptive alpha by state count following the change (and fix) of the learning rates used at every learning step described above.