[FV-RL] Improvements in documentation and messages shown mostly in simulators.py, and a small bug fix by storing job class rates as list
The files whose documentation and messages were improved are in:
- simulators.py
- QueuemM class
- and queues.AgeQueue class

On the other hand, the job_class_rates attribute in EnvQueueSingleBufferWithJobClasses is now checked that it's a list or a numpy array and always stored as a list.

Reason: I once passed this parameter job_class_rates as a numpy array and then got a functionality error because the EnvQueueSingleBufferWithJobClasses.getJobClassRates() method returned a numpy array instead of a list, making the calculation of valid_rates in the generate_event() function defined in simulators.py be a list of all NaNs, which should never happen as valid_rates is the result of concatenating the job class arrival rates (always non-NaN) and the service rates of servers that have at least one job in the queue (which could be an empty list, but the important thing is that the arrival rates and the service rates should be concatenated as LISTS, not as numpy arrays!).