[FV-RL] The reward returned by compute_reward_for_buffer_size() is 1.0 when no reward function is defined
Goal: Be able to run the FV estimation of the blocking probability without defining a reward function (as it is not needed). The logic behind returning 1.0 as reward when no reward function is defined has to do with the fact that the expected reward ni that case would be equal to the probability that the system's buffer size is at the buffer size passed to the function.