[Episodic Learners] The state value V(s) is always reset to 0 on TERMINAL states, regardless of the reset method
Summary: the ValueFunctionApprox class defined in value_functions.py accepts different methods to reset the state value function, namely "all zeros" or different flavours of random reset. For the random reset, however, we should always reset the state values to 0 for terminal states, which is how the value of terminal states is defined.

DETAILS:
- value_functions.py: the function approximation implementation now receives a new parameter in the constructor, `terminal_states`, which is used to set the value of terminal states to 0 (which is so by definition), regardless of the initial reset values of the value function.

- mc.py, td.py: adapted the call to the ValueFunctionApprox() class by passing the new required parameter `terminal_states`.