[MC] The MC learner now reads the value of the end-of-episode state in case the episode does not end at terminal state
All unit tests in test_MC_lambda.py pass.

Summary: In situations where the episode ends before reaching a terminal state it is sensible to use the value of the end state as an estimator of the return that would come until we reached a terminal state.
Example where this is useful: the Mountain Car environment where we establish a maximum number of steps for the episode which is commonly reached.

In addition, I updated the use of self._states to self.states because the latter has the last visited state added to the trajectory, whereas the former does not. This is important so that we can initialize the return G in learn_mc() to the value of the end state as self.V.getValue(self.state[-1]).