[utils] Redefined the basic.discretize() method to call np.digitize()
Goal: I observed bugs with my discretize() method where the assigned discretized interval was not the correct one when the value was the left bound of an interval (e.g. 1.2, when the left bound = 1.2, i.e. not even with the potential issue of rounding decimal --but probably the error comes when I divide by dx inside discretize(), which may not always generate the same decimals...?)

My previous discretize() function was renamed to deprecated_discretize().

I added new tests on left bounds of intervals as input.