[House] Removed token detected by GitGuardian
There was an access token in a URL that refers to the inp-toulouse website from where I had downloaded a few papers that was detected by GitGuardian as a potential security threat when I made the RL repository public in GitHub (in order to share it with Mingliang Yao, the intern at IRIT who started in Jun-2022).

I also improved a couple of descriptions and comments.