[FV] FV Speed Up #1: shortened data frame with Phi values
The execution time of the FV unit tests in test_estimators.py went down by half: the execution time of the 3 tests in test_estimators.py run by test_fv_implementation_single_server() went from ~60 sec down to ~30 sec.

This was thanked to following reduction of execution times of functions in run_simulation_fv() in simulators.py:
- merge_proba_survival_and_phi() went from ~45 msec down to 0.5 msec.
- estimate_proba_stationary() went from ~45 sec down to 0.002 sec.

DETAILS:
- simulators.py: the functions defined in run_simulation_fv() now only add a new row to the Phi(t,bs) when Phi(t,bs) changes w.r.t. its previous value, i.e. to its value prior to the latest event observed at time t. This reduced enormously the size of the Phi(t,bs) array being merged with the P(T>t) array, as well as the number of products to perform and elements to add when computing the integral in the numerator of the FV estimator.
Note also that the value of Phi(t,bs) are now stored in different dictionary entries (instead of in the same data frame as before) because the number of rows in the Phi(t,bs) array may vary for each bs.