[FV] FV Speed Up #2: the empirical mean behind Phi is incrementally updated
The 3 unit tests on the FV estimation in test_estimators.py pass.

The execution time based on test_QB.py on K=40, N=400 particles,T=200*N, J=0.5K didn't change much (e.g. from 93 secs to 90 secs for the FV part of the estimation process), but this is not a significant change.

DETAILS:
- simulators.py: the function update_phi() defined in run_simulation_fv() was split into two functions: initialize_phi(), which computes the empirical mean for the first time, and update_phi() which UPDATES the empirical mean from its previous value, based on the knowledge of the buffer size change in the particle that just changed state at t.